feat: extract shared gateway ensure crate
Extract ~400 lines of duplicated gateway auto-launch primitives from
dcc-mcp-cli and dcc-mcp-sidecar into a new dcc-mcp-gateway-ensure crate.

Shared primitives:
- Health check probes (/health endpoint)
- File-based launch lock with stale reclaim
- Detached process spawn (cross-platform)
- Gateway-ready polling with configurable timeout
- PID file read/write/remove
- Cross-platform process liveness check and termination
- Default registry directory resolution
- Gateway command-line argument construction

Consumers (CLI and sidecar) keep their unique orchestrator logic:
- CLI: lock-loser waiting path, gateway binary discovery, pidfile management
- Sidecar: version takeover with ElectionInfo/FileRegistry sentinel

Also unifies is_process_alive to a subprocess-based implementation
(replacing sidecar's sysinfo-based version).

Refs: PIP-1418
