+++ crates/dcc-mcp-cli/Cargo.toml
@@ -16,6 +16,7 @@ path = "src/main.rs"
 anyhow.workspace = true
 clap = { version = "4", features = ["derive", "env"] }
 dcc-mcp-catalog.workspace = true
+dcc-mcp-gateway-ensure.workspace = true
 dcc-mcp-marketplace.workspace = true
 dcc-mcp-skills.workspace = true
 dirs.workspace = true
