+++ crates/dcc-mcp-sidecar/src/gateway_daemon/launcher.rs
@@ -1,37 +1,13 @@
-use std::ffi::OsString;
-use std::fs::{File, OpenOptions};
-use std::path::{Path, PathBuf};
-use std::process::{Command, Stdio};
-use std::time::{Duration, Instant};
+use std::path::PathBuf;
+use std::time::Duration;
 
 use anyhow::Context as _;
 use dcc_mcp_gateway::{ElectionInfo, has_newer_sentinel, is_newer_election};
+use dcc_mcp_gateway_ensure as ensure;
 use dcc_mcp_transport::discovery::file_registry::FileRegistry;
 use dcc_mcp_transport::discovery::types::{GATEWAY_SENTINEL_DCC_TYPE, ServiceEntry};
 
-const ENV_GATEWAY_LAUNCH_LOCK_STALE_SECS: &str = "DCC_MCP_GATEWAY_LAUNCH_LOCK_STALE_SECS";
-const DEFAULT_GATEWAY_LAUNCH_LOCK_STALE_SECS: u64 = 30;
 const GATEWAY_SENTINEL_STALE_SECS: u64 = 30;
-/// Env var that overrides the gateway-ensure wait timeout.
-const ENV_GATEWAY_ENSURE_TIMEOUT_SECS: &str = "DCC_MCP_GATEWAY_ENSURE_TIMEOUT_SECS";
-/// Default ensure-wait timeout (15 s) when the env var is not set.
-const DEFAULT_ENSURE_TIMEOUT_SECS: u64 = 15;
-
-/// Resolve the effective ensure-wait timeout: env var
-/// `DCC_MCP_GATEWAY_ENSURE_TIMEOUT_SECS` > explicit arg > crate default (15 s).
-fn resolve_ensure_timeout_secs(explicit_secs: u64) -> u64 {
-    std::env::var(ENV_GATEWAY_ENSURE_TIMEOUT_SECS)
-        .ok()
-        .and_then(|raw| raw.parse::<u64>().ok())
-        .filter(|secs| *secs > 0)
-        .unwrap_or({
-            if explicit_secs > 0 {
-                explicit_secs
-            } else {
-                DEFAULT_ENSURE_TIMEOUT_SECS
-            }
-        })
-}
 
 /// Helpers for auto-launching the standalone gateway from inside another
 /// process (the per-DCC sidecar / embedded server).
@@ -65,7 +41,7 @@ pub async fn ensure_gateway_running(opts: &EnsureGatewayOptions) -> anyhow::Resu
         return Ok(());
     }
 
-    if gateway_health_ok(&opts.host, opts.port).await {
+    if ensure::gateway_health_ok(&opts.host, opts.port).await {
         try_version_takeover(opts).await?;
         return Ok(());
     }
@@ -73,13 +49,13 @@ pub async fn ensure_gateway_running(opts: &EnsureGatewayOptions) -> anyhow::Resu
     std::fs::create_dir_all(&opts.registry_dir)
         .with_context(|| format!("creating registry dir {}", opts.registry_dir.display()))?;
     let lock_path = opts.registry_dir.join("gateway-launch.lock");
-    match acquire_launch_lock(&lock_path) {
+    match ensure::acquire_launch_lock(&lock_path) {
         Ok(_lock) => {
-            if gateway_health_ok(&opts.host, opts.port).await {
+            if ensure::gateway_health_ok(&opts.host, opts.port).await {
                 try_version_takeover(opts).await?;
                 return Ok(());
             }
-            spawn_detached_gateway(opts)?;
+            spawn_detached_gateway_now(opts)?;
         }
         Err(err) if err.kind() == std::io::ErrorKind::AlreadyExists => {
             tracing::info!(
@@ -90,13 +66,15 @@ pub async fn ensure_gateway_running(opts: &EnsureGatewayOptions) -> anyhow::Resu
         Err(err) => return Err(err).with_context(|| format!("creating {}", lock_path.display())),
     }
 
-    wait_gateway_ready(
+    ensure::wait_gateway_ready(
         &opts.host,
         opts.port,
-        Duration::from_secs(resolve_ensure_timeout_secs(0)),
+        Duration::from_secs(ensure::resolve_ensure_timeout_secs(0)),
     )
     .await
 }
+
+// ── Version takeover ───────────────────────────────────────────────────────
 
 /// When the gateway port is already occupied, check whether we carry a newer
 /// version than the running gateway.  If we do, write a sentinel entry with
@@ -171,9 +149,9 @@ async fn try_version_takeover(opts: &EnsureGatewayOptions) -> anyhow::Result<()>
 
     // Wait for the old gateway to yield (up to ~20 s for the 15 s cleanup
     // interval + grace).
-    let deadline = Instant::now() + Duration::from_secs(20);
-    while Instant::now() < deadline {
-        if !gateway_health_ok(&opts.host, opts.port).await {
+    let deadline = tokio::time::Instant::now() + Duration::from_secs(20);
+    while tokio::time::Instant::now() < deadline {
+        if !ensure::gateway_health_ok(&opts.host, opts.port).await {
             tracing::info!("old gateway yielded — spawning new gateway");
             return spawn_gateway_with_lock(opts).await;
         }
@@ -190,12 +168,12 @@ async fn spawn_gateway_with_lock(opts: &EnsureGatewayOptions) -> anyhow::Result<
     std::fs::create_dir_all(&opts.registry_dir)
         .with_context(|| format!("creating registry dir {}", opts.registry_dir.display()))?;
     let lock_path = opts.registry_dir.join("gateway-launch.lock");
-    match acquire_launch_lock(&lock_path) {
+    match ensure::acquire_launch_lock(&lock_path) {
         Ok(_lock) => {
-            if gateway_health_ok(&opts.host, opts.port).await {
+            if ensure::gateway_health_ok(&opts.host, opts.port).await {
                 return Ok(());
             }
-            spawn_detached_gateway(opts)?;
+            spawn_detached_gateway_now(opts)?;
         }
         Err(err) if err.kind() == std::io::ErrorKind::AlreadyExists => {
             tracing::info!(
@@ -205,170 +183,35 @@ async fn spawn_gateway_with_lock(opts: &EnsureGatewayOptions) -> anyhow::Result<
         }
         Err(err) => return Err(err).with_context(|| format!("creating {}", lock_path.display())),
     }
-    wait_gateway_ready(
+    ensure::wait_gateway_ready(
         &opts.host,
         opts.port,
-        Duration::from_secs(resolve_ensure_timeout_secs(0)),
+        Duration::from_secs(ensure::resolve_ensure_timeout_secs(0)),
     )
     .await
 }
 
-async fn wait_gateway_ready(host: &str, port: u16, timeout: Duration) -> anyhow::Result<()> {
-    let deadline = Instant::now() + timeout;
-    while Instant::now() < deadline {
-        if gateway_health_ok(host, port).await {
-            return Ok(());
-        }
-        tokio::time::sleep(Duration::from_millis(150)).await;
-    }
-    anyhow::bail!(
-        "gateway did not become healthy at http://{host}:{port}/health within {timeout:?}"
-    )
-}
+// ── Spawn helper ───────────────────────────────────────────────────────────
 
-pub(super) async fn gateway_health_ok(host: &str, port: u16) -> bool {
-    gateway_health_ok_with_timeout(host, port, Duration::from_millis(600)).await
-}
-
-pub(super) async fn gateway_health_ok_with_timeout(
-    host: &str,
-    port: u16,
-    timeout: Duration,
-) -> bool {
-    let url = format!("http://{host}:{port}/health");
-    let client = match reqwest::Client::builder().timeout(timeout).build() {
-        Ok(client) => client,
-        Err(_) => return false,
-    };
-    client
-        .get(url)
-        .send()
-        .await
-        .is_ok_and(|resp| resp.status().is_success())
-}
-
-struct LaunchLock {
-    _file: File,
-    path: PathBuf,
-}
-
-impl Drop for LaunchLock {
-    fn drop(&mut self) {
-        let _ = std::fs::remove_file(&self.path);
-    }
-}
-
-fn acquire_launch_lock(path: &Path) -> std::io::Result<LaunchLock> {
-    acquire_launch_lock_with_stale(path, gateway_launch_lock_stale_after())
-}
-
-fn acquire_launch_lock_with_stale(
-    path: &Path,
-    stale_after: Duration,
-) -> std::io::Result<LaunchLock> {
-    match create_launch_lock(path) {
-        Ok(lock) => Ok(lock),
-        Err(err) if err.kind() == std::io::ErrorKind::AlreadyExists => {
-            if remove_stale_launch_lock(path, stale_after)? {
-                create_launch_lock(path)
-            } else {
-                Err(err)
-            }
-        }
-        Err(err) => Err(err),
-    }
-}
-
-fn create_launch_lock(path: &Path) -> std::io::Result<LaunchLock> {
-    OpenOptions::new()
-        .write(true)
-        .create_new(true)
-        .open(path)
-        .map(|file| LaunchLock {
-            _file: file,
-            path: path.to_path_buf(),
-        })
-}
-
-fn gateway_launch_lock_stale_after() -> Duration {
-    std::env::var(ENV_GATEWAY_LAUNCH_LOCK_STALE_SECS)
-        .ok()
-        .and_then(|raw| raw.parse::<u64>().ok())
-        .filter(|secs| *secs > 0)
-        .map(Duration::from_secs)
-        .unwrap_or_else(|| Duration::from_secs(DEFAULT_GATEWAY_LAUNCH_LOCK_STALE_SECS))
-}
-
-fn remove_stale_launch_lock(path: &Path, stale_after: Duration) -> std::io::Result<bool> {
-    if !launch_lock_is_stale(path, stale_after)? {
-        return Ok(false);
-    }
-
-    // Re-check immediately before unlinking so a fresh lock created by another
-    // sidecar is not removed after the stale check wins a race.
-    if !launch_lock_is_stale(path, stale_after)? {
-        return Ok(false);
-    }
-
-    match std::fs::remove_file(path) {
-        Ok(()) => Ok(true),
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => Ok(true),
-        Err(err) => Err(err),
-    }
-}
-
-fn launch_lock_is_stale(path: &Path, stale_after: Duration) -> std::io::Result<bool> {
-    let modified = match std::fs::metadata(path) {
-        Ok(meta) => meta.modified()?,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(true),
-        Err(err) => return Err(err),
-    };
-    let age = modified.elapsed().unwrap_or_default();
-    Ok(age >= stale_after)
-}
-
-fn spawn_detached_gateway(opts: &EnsureGatewayOptions) -> anyhow::Result<()> {
-    let exe = std::env::current_exe().context("resolving current executable")?;
-    let mut cmd = Command::new(exe);
-    cmd.args(gateway_command_args(opts))
-        .env("DCC_MCP_REGISTRY_DIR", &opts.registry_dir)
-        .stdin(Stdio::null())
-        .stdout(Stdio::null())
-        .stderr(Stdio::null());
-
-    #[cfg(windows)]
-    {
-        use std::os::windows::process::CommandExt;
-        const CREATE_NO_WINDOW: u32 = 0x0800_0000;
-        const DETACHED_PROCESS: u32 = 0x0000_0008;
-        cmd.creation_flags(CREATE_NO_WINDOW | DETACHED_PROCESS);
-    }
-
-    cmd.spawn().context("spawning standalone gateway")?;
+fn spawn_detached_gateway_now(opts: &EnsureGatewayOptions) -> anyhow::Result<()> {
+    let exe =
+        std::env::current_exe().context("resolving current executable for detached gateway")?;
+    let cmd_args = ensure::gateway_command_args(
+        &opts.host,
+        opts.port,
+        opts.name.as_deref(),
+        &opts.remote_host,
+        opts.remote_port,
+        opts.gateway_idle_timeout_secs,
+    );
+    ensure::spawn_detached_gateway(&exe, &cmd_args, &opts.registry_dir)?;
     tracing::info!(port = opts.port, "spawned standalone gateway process");
     Ok(())
 }
 
-fn gateway_command_args(opts: &EnsureGatewayOptions) -> Vec<OsString> {
-    let mut args = vec![
-        OsString::from("gateway"),
-        OsString::from("--host"),
-        OsString::from(&opts.host),
-        OsString::from("--port"),
-        OsString::from(opts.port.to_string()),
-        OsString::from("--remote-host"),
-        OsString::from(&opts.remote_host),
-        OsString::from("--remote-port"),
-        OsString::from(opts.remote_port.to_string()),
-    ];
-    if let Some(name) = opts.name.as_ref().filter(|name| !name.trim().is_empty()) {
-        args.push(OsString::from("--name"));
-        args.push(OsString::from(name));
-    }
-    args.push(OsString::from("--gateway-idle-timeout-secs"));
-    args.push(OsString::from(opts.gateway_idle_timeout_secs.to_string()));
-    args
-}
+// ── Re-exports ─────────────────────────────────────────────────────────────
+
+pub use ensure::gateway_health_ok_with_timeout;
 
 #[cfg(test)]
 mod tests {
@@ -376,30 +219,19 @@ mod tests {
 
     #[test]
     fn auto_launch_gateway_args_do_not_include_registry_dir_flag() {
-        let opts = EnsureGatewayOptions {
-            host: "127.0.0.1".to_string(),
-            port: 9765,
-            name: Some("gateway-for-test".to_string()),
-            registry_dir: PathBuf::from(r"C:\tmp\dcc-mcp-registry"),
-            remote_host: "0.0.0.0".to_string(),
-            remote_port: 59765,
-            crate_version: None,
-            adapter_version: None,
-            adapter_dcc: None,
-            gateway_idle_timeout_secs: 30,
-        };
-
-        let args: Vec<String> = gateway_command_args(&opts)
-            .into_iter()
-            .map(|arg| arg.to_string_lossy().to_string())
-            .collect();
+        let argv: Vec<String> = ensure::gateway_command_args(
+            "127.0.0.1", 9765, Some("gateway-for-test"), "0.0.0.0", 59765, 30,
+        )
+        .into_iter()
+        .map(|arg| arg.to_string_lossy().to_string())
+        .collect();
 
         assert!(
-            !args.iter().any(|arg| arg == "--registry-dir"),
+            !argv.iter().any(|arg| arg == "--registry-dir"),
             "auto-launched gateway should inherit DCC_MCP_REGISTRY_DIR instead of exposing --registry-dir in the command line"
         );
-        assert!(args.iter().any(|arg| arg == "gateway"));
-        assert!(args.iter().any(|arg| arg == "--name"));
+        assert!(argv.iter().any(|arg| arg == "gateway"));
+        assert!(argv.iter().any(|arg| arg == "--name"));
     }
 
     #[test]
@@ -408,7 +240,7 @@ mod tests {
         let path = dir.path().join("gateway-launch.lock");
         std::fs::write(&path, "stale").unwrap();
 
-        let lock = acquire_launch_lock_with_stale(&path, Duration::ZERO)
+        let lock = ensure::acquire_launch_lock_with_stale(&path, Duration::ZERO)
             .expect("stale launch lock should be reclaimed");
 
         assert!(path.exists());
@@ -422,7 +254,7 @@ mod tests {
         let path = dir.path().join("gateway-launch.lock");
         std::fs::write(&path, "busy").unwrap();
 
-        let err = match acquire_launch_lock_with_stale(&path, Duration::from_secs(3600)) {
+        let err = match ensure::acquire_launch_lock_with_stale(&path, Duration::from_secs(3600)) {
             Ok(_) => panic!("fresh launch lock should remain busy"),
             Err(err) => err,
         };
